Redress undesired adaptor type changes
The initial overload of serial left these undesired type changes in
place. Return to initial, valid, state.

+++ lib/ecto/adapters/postgres/connection.ex
@@ -844,7 +844,7 @@ if Code.ensure_loaded?(Postgrex) do
     defp reference_name(%Reference{name: name}, _table, _column),
       do: quote_name(name)
 
-    defp reference_column_type(:serial, _opts), do: "bigint"
+    defp reference_column_type(:serial, _opts), do: "integer"
     defp reference_column_type(:bigserial, _opts), do: "bigint"
     defp reference_column_type(type, opts), do: column_type(type, opts)
